src/requirements.txt
PyJWT==2.1.0
cryptography
